typescript support and better build/release pipeline. Probably support other environments besides node with those changes
