+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# abacatepay-nodejs-sdk
+
+## 1.0.1
+
+### Patch Changes
+
+- This adds better typescript support, prevents of importing package.json and better build pipeline overall
