+++ package.json
@@ -1,16 +1,37 @@
 {
   "name": "abacatepay-nodejs-sdk",
-  "version": "1.0.0",
-  "main": "src/index.js",
-  "types": "types.d.ts",
+  "files": [
+    "dist",
+    "README",
+    "LICENSE",
+    "package.json"
+  ],
+  "version": "1.0.1",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "define:change": "changeset",
+    "release:apply": "changeset version",
+    "release:publish": "changeset publish",
+    "prebuild": "node -p \"'/*This file is auto generated during build, DO NOT CHANGE OR MODIFY */\\n\\nexport const ABACATE_PAY_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.ts",
+    "build": "tsup ./src --out-dir ./dist --format cjs,esm --dts --silent --no-splitting --no-splitting --clean"
   },
-  "keywords": [],
+  "keywords": [
+    "abacate",
+    "abacatepay",
+    "sdk"
+  ],
   "author": "Christopher Ribeiro <christo_campiglia@hotmail.com>",
   "license": "MIT",
   "description": "AbacatePay NodeJS SDK",
   "dependencies": {
     "node-fetch": "^3.3.2"
+  },
+  "devDependencies": {
+    "@changesets/cli": "^2.27.10",
+    "tsup": "^8.3.5",
+    "typescript": "^5.7.2"
   }
 }
