+++ src/version.ts
@@ -1,0 +1,3 @@
+/*This file is auto generated during build, DO NOT CHANGE OR MODIFY */
+
+export const ABACATE_PAY_VERSION = "1.0.0";
